feat(配置/UI): 将4mA和20mA量程类型从uint32改为float
修改配置参数和UI界面以支持浮点数量程，包括：
- 修改全局配置结构和默认值
- 更新配置获取/设置函数接口
- 添加Modbus浮点数寄存器支持
- 扩展UI设置项结构支持浮点数路径
- 实现浮点数编辑和显示逻辑

+++ App/app_config.h
@@ -68,11 +68,11 @@ void app_config_set_calibration_4ma(uint32_t value);
 uint32_t app_config_get_calibration_20ma(void);
 void app_config_set_calibration_20ma(uint32_t value);
 
-uint32_t app_config_get_range_4ma(void);
-void app_config_set_range_4ma(uint32_t value);
+float app_config_get_range_4ma(void);
+void app_config_set_range_4ma(float value);
 
-uint32_t app_config_get_range_20ma(void);
-void app_config_set_range_20ma(uint32_t value);
+float app_config_get_range_20ma(void);
+void app_config_set_range_20ma(float value);
 
 uint32_t app_config_get_point_num(void);
 void app_config_set_point_num(uint32_t value);
